Create User entity, and map tasks to users
We would use an Auth mechanism to know the user information, and tie
Tasks to that User. Until then, we use a custom header to pass the user
ID along.

Creating a Task now requires a User, and will perform a look up of the
user. Listing Tasks now only lists the tasks that belong to that user.
later, we can build auth to restrict the get Task API to only return a
Task if it belongs to the given user.

Modified unit tests to create Tasks for unique users on a per test basis
to provide isolation between the tests.