Refactor the Post HTMX responses to not use 302 redirect
Although HTMX can seamlessly do the 302 post-redirect-get loop without a
reload, you don't have to do this by default. I refactored the update
endpoint to return the re-rendered Task for the updated task. For the
re-rank endpoint, it is set to do nothing at all by returning a 204 (no
content).