Improve UI for editing task title; fix bug with sorting
To improve the UI for editing the task title, it now only triggers when
double clicking over the title itself. I also added feather icons, and
included icons indicating save/cancel (in addition to being able to
press ENTER/ESC).

While working on the form, I discovered a bug with the sortable feature.
I had followed the example online, which suggested using a form
attribute. However, I also need forms in the Tasks themselves, which
would lead to nested forms, which you cannot really do in HTML. To solve
this, I used a div instead of a form, and used some HTMX magic to still
capture the input data to send to the backend.