Extend the list Tasks endpoints to allow searching by state
By default, the application does not show completed Tasks, because users
should be able to focus on the things they still need to get done.
However, they may still want to review Tasks they have already
completed.

This change introduces a toggle switch in the UI to view completed
tasks. Completed tasks are shown at the bottom of the list of Tasks, and
they are not sortable.

To implement the change, we extend the list Task APIs to take in a
"state" query string, and filter on any valid states the user provide.