Improve UX for displaying completed Tasks
PROBLEM
The UX was a bit janky. The toggle did not smoothly transition, because
the whole page would be replaced when switching between views. The tasks
list included all open/complete tasks, and used sorting to control
their position; although the completed tasks could not be re-ranked
using the drag and drop feature, the open tasks could be dragged among
the complete tasks, which was weird.

There were also some issues with the implementation details. Passing the
showCompleted header to the backend seemed janky. The sorting in the API
was coupled to the UX use case. It was slightly inefficient to search
for all open/complete tasks from the UI, when the user would already
have open tasks visible.

SOLUTION
I split up the template so listing tasks would only return the HTML for
the list. That way, it could return separate lists for open and
complete. The view now has two distinct lists. Searching for the
completed tasks no longer requires re-rendering the whole page
(including open tasks), which makes it smoother and more efficient. With
the completed tasks in a separate DOM node, it's no longer possible to
drag/drop open tasks into the completed tasks.