Update Task-to-User to use Lazy fetching
By default, Hibernate association annotations use an Eager fetch
strategy. Operations on Tasks do not currently require User data (we
only search by the user id). Conversely, operations on the User do not
require Tasks. And yet, with eager fetching, that data is eagerly
fetched from the database. And it could be executed with an N+1 query.
I'm not worried about performance now for a prototype, but I would like
to learn about and adopt best practices as early as possible to build in
the habit.

To fix this issue, I set the associations to be lazy. All tests continue
to pass, and everything is still working on the UI.