Implement feature to re-open tasks
Users may want to re-open completed tasks for various reasons. Perhaps
they completed the task by mistake, or else they thought the task was
finished but in fact there is more work to be done.

This commit implements the full feature (backend + frontend).

Because alpineJS is a critical dependency of the frondend experience
(its usefulness has been proved beyond a doubt), it has been added as a
webjar instead of a CDN. This is also useful for local development when
the CDN is taking a long time to respond, or due to network issues.