Improve UI so completing task with "show completed" on appends completed task to list
PROBLEM
If a user has the "show completed" toggle switched on, and if they
complete some tasks, then those tasks vanish from the DOM. In order to
see the tasks, they have to switch the toggle off/on to refresh.

SOLUTION
Re-use the pattern of using the hx-swap-oob to clear the now completed
task from the "open task" node, and append the completed task HTMl to
the "completed task" node.