Modify the root "/" url to direct to the task list app
This prototype really has only one active page at "/my-tasks", and I
don't have a plan to add a home page (and I don't want to keep Quarkus's
default homepage), so this does a 307 redirect.

In addition, this commit adds some documentation to the README, and
additional Javadoc to make commenting a bit more consistent.